==> analysis/analysis_old/01_Wrangle_data_PMX_code.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx", low_memory=False)

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_pmx_00_19_lw = df_pmx_00_19["estado", "municipio"].str.lower()

df_pmx_00_19_lw = df_pmx_00_19["estado"].str.lower()

df_pmx_00_19_lw = df_pmx_00_19.str.lower()

df_pmx_00_19_lw = df_pmx_00_19[].str.lower()

df_pmx_00_19_lw = df_pmx_00_19["year", "estado"].str.lower()

df_pmx_00_19["estado"] = df_pmx_00_19["estado"].str.lower()

df_pmx_00_19["estado"] = df_pmx_00_19["estado"].str.lower()
df_pmx_00_19["municipio"] = df_pmx_00_19["municipio"].str.lower()

df_pmx_00_19_lw["estado"] = df_pmx_00_19["estado"].str.lower()
df_pmx_00_19_lw["municipio"] = df_pmx_00_19["municipio"].str.lower()

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower()
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower()

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower()
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower()
df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.rstrip()
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.rstrip()

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower()
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower()
df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.replace(' ', '')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(' ', '')

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower(df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.replace(' ', ''))
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower(df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(' ', ''))

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower(.str.replace(' ', ''))
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower(.str.replace(' ', ''))

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '')

def remove_accents(a):
    return unidecode.unidecode(a.decode('utf-8'))

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').apply(remove_accents)
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').apply(remove_accents)



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')

df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw.assign(id_mun = estado + municipio)


df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw.assign(id_mun = NOM_ENT + NOM_MUN)



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw.assign(id_mun = "estado" + "municipio")


df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw.assign(id_mun = "NOM_ENT" + "NOM_MUN")



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw.assign(id_mun = "estado" + "municipio")


df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw = df_codes_mx_lw.assign(id_mun = "NOM_ENT" + "NOM_MUN")



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw.assign(id_mun = "estado" + "municipio")


df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw = df_codes_mx_lw.assign(id_mun = "NOM_ENT" + "NOM_MUN")
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_codes_mx_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_codes_mx_lw = df_codes_mx_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_codes_mx_lw = df_codes_mx_lw.replace("Coahuila de Zaragoza", "Coahuila")
df_codes_mx_lw = df_codes_mx_lw.replace("Michoacán de Ocampo", "Michoacán")
df_codes_mx_lw = df_codes_mx_lw.replace("Veracruz de Ignacio de la Llave", "Veracruz")

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_codes_mx_lw = df_codes_mx_lw.replace("Coahuila de Zaragoza", "Coahuila")
df_codes_mx_lw = df_codes_mx_lw.replace("Michoacán de Ocampo", "Michoacán")
df_codes_mx_lw = df_codes_mx_lw.replace("Veracruz de Ignacio de la Llave", "Veracruz")

df_pmx_00_19_lw["municipio"].str.replace("Delegación", "")
df_pmx_00_19_lw["municipio"].str.replace("Deleg. ", "")


            
                    
df_pmx_00_19_lw["A"].str.replace("a","")

                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_codes_mx_lw = df_codes_mx_lw.replace("Coahuila de Zaragoza", "Coahuila")
df_codes_mx_lw = df_codes_mx_lw.replace("Michoacán de Ocampo", "Michoacán")
df_codes_mx_lw = df_codes_mx_lw.replace("Veracruz de Ignacio de la Llave", "Veracruz")

df_pmx_00_19_lw["municipio"].str.replace("Delegación", "")
df_pmx_00_19_lw["municipio"].str.replace("Deleg. ", "")

                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

df_codes_mx_lw = df_codes_mx_lw.replace("Coahuila de Zaragoza", "Coahuila")
df_codes_mx_lw = df_codes_mx_lw.replace("Michoacán de Ocampo", "Michoacán")
df_codes_mx_lw = df_codes_mx_lw.replace("Veracruz de Ignacio de la Llave", "Veracruz")

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace("Delegación", "")
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace("Deleg. ", "")

                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace("Delegación", "")
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace("Deleg. ", "")

                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(dict_replacers_string)

                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].replace(dict_replacers_string)


                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].replace(dict_replacers_string)


                

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].replace(dict_replacers_string)

for k, v in dict_replacers_string.items():
    df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(k, v)               

            

            

df_pmx_00_19 = pd.read_excel("data-raw/PMX_00_19_Serendipia.xlsx")

df_codes_mx = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_pmx_00_19_lw = df_pmx_00_19

df_codes_mx_lw = df_codes_mx

dict_replacers_code = {"Coahuila de Zaragoza":"Coahuila","Michoacán de Ocampo":"Michoacán",
                       "Veracruz de Ignacio de la Llave":"Veracruz"} 

df_codes_mx_lw = df_codes_mx_lw.replace(dict_replacers_code)

dict_replacers_string = {"Delegación":"","Deleg. ":"","Del. ":""} 

df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].replace(dict_replacers_string)

for k, v in dict_replacers_string.items():
    df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.replace(k, v)               


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


summary.df_join_pipe_code['_merge']


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


sum.df_join_pipe_code['_merge']


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


describe.df_join_pipe_code['_merge']


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


df_join_pipe_code['_merge'].summary


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


df_join_pipe_code['_merge'].summary()


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


df_join_pipe_code.summary()


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)


df_join_pipe_code['_merge'].describe()


df_pmx_00_19_lw = df_pmx_00_19_lw.replace("Coahuila de Zaragoza", "Coahuila")

df_pmx_00_19_lw["estado"] = df_pmx_00_19_lw["estado"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw["municipio"] = df_pmx_00_19_lw["municipio"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_pmx_00_19_lw['id_mun'] = df_pmx_00_19_lw.estado + df_pmx_00_19_lw.municipio



df_codes_mx_lw["NOM_ENT"] = df_codes_mx_lw["NOM_ENT"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw["NOM_MUN"] = df_codes_mx_lw["NOM_MUN"].str.lower().str.replace(' ', '').str.normalize('NFKD').str.encode('ascii',errors='ignore').str.decode('utf-8')
df_codes_mx_lw['id_mun'] = df_codes_mx_lw.NOM_ENT + df_codes_mx_lw.NOM_MUN



df_join_pipe_code = pd.merge(df_pmx_00_19_lw,
                             df_codes_mx_lw,
                             on='id_mun',
                             how='outer',
                             indicator=True)

df_join_pipe_code['codigo inegi'] = df_join_pipe_code['codigo inegi'].str.zfill(2)

class(df_join_pipe_code['codigo inegi'] = df_join_pipe_code['codigo inegi'])

class(df_join_pipe_code['codigo inegi'])

type(df_join_pipe_code['codigo inegi'])

df_join_pipe_code.dtypes

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

  %notebook "01_Wrangle_data_PMX_code.ipynb"     

  %notebook "analysis/01_Wrangle_data_PMX_code.ipynb"     

  %notebook "analysis/"

  %notebook "analysis/01_Wrangle_data_PMX_code.ipynb"

  %notebook ""

  %notebook "analysis/mx.ipynb"

  %notebook analysis/mx.ipynb
