==> analysis/analysis_old/05_wrangle_pipelines_and_thefts.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_PMX_SP_15_19_CC_f = pd.read_parquet("data/df_PMX_SP_15_19_thefts_pp.parquet")

df_pipelines_mx = pd.read_parquet("data/df_pipelines_mx.parquet")



df_PMX_15_19_thefts_pp = pd.merge(df_pipelines_mx,
                             df_PMX_SP_15_19_CC_f,
                             on='CVEGEO',
                             how='outer',
                             indicator=True)


df_PMX_15_19_thefts_pp



df_PMX_15_19_thefts_pp.loc[df_PMX_15_19_thefts_pp["Num_thefts"] == 0, "pipeline"] = 1

df_PMX_15_19_thefts_pp["pipeline"].sum()


df_PMX_15_19_thefts_pp

df_PMX_15_19_thefts_pp.to_excel('data/df_PMX_15_19_thefts_pp_final.xlsx', index = False)
df_PMX_15_19_thefts_pp.to_parquet('data/df_PMX_15_19_thefts_pp_final.parquet')
