==> analysis/analysis_old/06_wrangle_homicides.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_cvegeo = pd.read_excel("data-raw/codigo municipio inegi.xlsx")

df_cvegeo = df_cvegeo[["codigo inegi"]].rename(columns = {"codigo inegi":"CVEGEO"})


df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].astype(str).str.zfill(7).str.replace('.',"")


df_cvegeo['CVEGEO'] = df_cvegeo['CVEGEO'].str[:-1]






df_cvegeo



df_homicides_15 = pd.read_excel("data-raw/INEGI_homic_2015.xls")
df_homicides_16 = pd.read_excel("data-raw/INEGI_homic_2016.xls")
df_homicides_17 = pd.read_excel("data-raw/INEGI_homic_2017.xls")
df_homicides_18 = pd.read_excel("data-raw/INEGI_homic_2018.xls")
df_homicides_19 = pd.read_excel("data-raw/INEGI_homic_2019.xls")
df_homicides_15_w = df_homicides_15

df_homicides_15.head(20)

dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
              "Unnamed: 2":"homic_15"} 


df_homicides_15_w = df_homicides_15[df_homicides_15['Unnamed: 1'].notnull()]

df_homicides_15_w = df_homicides_15_w.rename(columns = dict_var_homic)

df_homicides_15_w_var = df_homicides_15_w[["CVEGEO", "homic_15"]]

df_homicides_15_w_var= df_homicides_15_w_var[df_homicides_15_w_var['CVEGEO'].notnull()]

df_homicides_15_w_var_f = df_homicides_15_w_var[(df_homicides_15_w_var['CVEGEO'].str.len()>=3)]

df_homicides_15_w_var_f['homic_15']= df_homicides_15_w_var_f['homic_15'].str.replace(',',"")

df_homicides_15_w_var_f['CVEGEO']= df_homicides_15_w_var_f['CVEGEO'].str.replace(' ',"")

df_homicides_15_w_var_f['homic_15']= pd.to_numeric(df_homicides_15_w_var_f['homic_15'])

df_homicides_15_w_var_f

df_homicides_16.head(20)

dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
              "Unnamed: 2":"homic_16"} 


df_homicides_16_w = df_homicides_16[df_homicides_16['Unnamed: 1'].notnull()]

df_homicides_16_w = df_homicides_16_w.rename(columns = dict_var_homic)

df_homicides_16_w_var = df_homicides_16_w[["CVEGEO", "homic_16"]]

df_homicides_16_w_var= df_homicides_16_w_var[df_homicides_16_w_var['CVEGEO'].notnull()]

df_homicides_16_w_var_f = df_homicides_16_w_var[(df_homicides_16_w_var['CVEGEO'].str.len()>=3)]

df_homicides_16_w_var_f['homic_16']= df_homicides_16_w_var_f['homic_16'].str.replace(',',"")

df_homicides_16_w_var_f['CVEGEO']= df_homicides_16_w_var_f['CVEGEO'].str.replace(' ',"")

df_homicides_16_w_var_f['homic_16']= pd.to_numeric(df_homicides_16_w_var_f['homic_16'])

df_homicides_16_w_var_f

df_homicides_17.head(20)

dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
              "Unnamed: 2":"homic_17"} 


df_homicides_17_w = df_homicides_17[df_homicides_17['Unnamed: 1'].notnull()]

df_homicides_17_w = df_homicides_17_w.rename(columns = dict_var_homic)

df_homicides_17_w_var = df_homicides_17_w[["CVEGEO", "homic_17"]]

df_homicides_17_w_var= df_homicides_17_w_var[df_homicides_17_w_var['CVEGEO'].notnull()]

df_homicides_17_w_var_f = df_homicides_17_w_var[(df_homicides_17_w_var['CVEGEO'].str.len()>=3)]

df_homicides_17_w_var_f['homic_17']= df_homicides_17_w_var_f['homic_17'].str.replace(',',"")

df_homicides_17_w_var_f['CVEGEO']= df_homicides_17_w_var_f['CVEGEO'].str.replace(' ',"")

df_homicides_17_w_var_f['homic_17']= pd.to_numeric(df_homicides_17_w_var_f['homic_17'])

df_homicides_17_w_var_f

df_homicides_18.head(20)

dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
              "Unnamed: 2":"homic_18"} 


df_homicides_18_w = df_homicides_18[df_homicides_18['Unnamed: 1'].notnull()]

df_homicides_18_w = df_homicides_18_w.rename(columns = dict_var_homic)

df_homicides_18_w_var = df_homicides_18_w[["CVEGEO", "homic_18"]]

df_homicides_18_w_var= df_homicides_18_w_var[df_homicides_18_w_var['CVEGEO'].notnull()]

df_homicides_18_w_var_f = df_homicides_18_w_var[(df_homicides_18_w_var['CVEGEO'].str.len()>=3)]

df_homicides_18_w_var_f['homic_18']= df_homicides_18_w_var_f['homic_18'].str.replace(',',"")

df_homicides_18_w_var_f['CVEGEO']= df_homicides_18_w_var_f['CVEGEO'].str.replace(' ',"")

df_homicides_18_w_var_f['homic_18']= pd.to_numeric(df_homicides_18_w_var_f['homic_18'])

df_homicides_18_w_var_f

df_homicides_19.head(20)

dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
              "Unnamed: 2":"homic_19"} 


df_homicides_19_w = df_homicides_19[df_homicides_19['Unnamed: 1'].notnull()]

df_homicides_19_w = df_homicides_19_w.rename(columns = dict_var_homic)

df_homicides_19_w_var = df_homicides_19_w[["CVEGEO", "homic_19"]]

df_homicides_19_w_var= df_homicides_19_w_var[df_homicides_19_w_var['CVEGEO'].notnull()]

df_homicides_19_w_var_f = df_homicides_19_w_var[(df_homicides_19_w_var['CVEGEO'].str.len()>=3)]

df_homicides_19_w_var_f['homic_19']= df_homicides_19_w_var_f['homic_19'].str.replace(',',"")

df_homicides_19_w_var_f['CVEGEO']= df_homicides_19_w_var_f['CVEGEO'].str.replace(' ',"")

df_homicides_19_w_var_f['homic_19']= pd.to_numeric(df_homicides_19_w_var_f['homic_19'])

df_homicides_19_w_var_f

#dataframes_homic = [df_homicides_15,
#                    df_homicides_16,
#                    df_homicides_17,
#                    df_homicides_18,
#                    df_homicides_19]
#
#list_of_df = []
#
#df_homicides_15.head(20)
#
#dict_var_homic = {"Defunciones por homicidios": "CVEGEO",
#              "Unnamed: 2":"homic"} 
#
#
#for dataframe in dataframes_homic:
#    
#   dataframe_final = dataframe[dataframe['Unnamed: 1'].notnull()]
#   
#   dataframe+"_w" = dataframe+"_w".rename(columns = dict_var_homic)
#   
#   dataframe+"_w_var" = dataframe+"_w"[["CVEGEO", "homic"]]
#   
#   dataframe+"_w_var"= dataframe+"_w_var"[dataframe+"_w_var"['CVEGEO'].notnull()]
#   
#   dataframe+"_w_var_f" = pdataframe+"_w_var")[(dataframe+"_w_var")['CVEGEO'].str.len()>=3)]
#   
#   dataframe+"_w_var_f"['homic']= dataframe+"_w_var_f")['homic'].str.replace(',',"")
#   
#   dataframe+"_w_var_f"['CVEGEO']= dataframe+"_w_var_f")['CVEGEO'].str.replace(' ',"")
#   
#   dataframe+"_w_var_f"['homic']= pd.to_numeric(dataframe +"_w_var_f"['homic'])
#   
#   list_of_df.append(dataframe+"_w_var_f")
# 

df_homicides_15.index

len(df_homicides_15)

df_homicies_15_19 = [(df_homicies_15_22.year <=2019)]


df_homicies_15_19 = df_homicies_15_22[df_homicies_15_22['Tipo de delito'].str.contains("Homicidio|Narcomenudeo|Robo|Secuestro|Trata de personas|Extorsion")]


df_homicies_15_19_homic = df_homicies_15_19[~df_homicies_15_19['Subtipo de delito'].str.contains("Homicidio culposo")]

df_homicies_15_19_homic["delitos"] = df_homicies_15_19_homic[["Enero", "Febrero", "Marzo", "Abril", "Mayo", "Junio", "Julio", "Agosto", "Septiembre", "Octubre", "Noviembre", "Diciembre"]].sum(axis=1)


df_homicies_15_19_homic.drop(["Enero", "Febrero", "Marzo", "Abril", "Mayo", "Junio", "Julio", "Agosto", "Septiembre", "Octubre", "Noviembre", "Diciembre", "year", "Clave_Ent"], axis = 1, inplace = True)

df_homicies_15_19_homic_group = df_homicies_15_19_homic.groupby(["Cve_Municipio", "Tipo de delito"]).mean().reset_index()






df_homicies_15_19_prueba = df_homicies_15_22[df_homicies_15_22['Subtipo de delito'].str.contains("Homicidio doloso")]

df_homicies_15_19_prueba["delitos"] = df_homicies_15_19_prueba[["Enero", "Febrero", "Marzo", "Abril", "Mayo", "Junio", "Julio", "Agosto", "Septiembre", "Octubre", "Noviembre", "Diciembre"]].sum(axis=1)


df_homicies_15 = df_homicies_15_19_prueba[(df_homicies_15_19_prueba.year == 2015)] 

df_homicies_15["delitos"].sum()

df_homicies_15_19_wide = df_homicies_15_19_homic_group.pivot(index='Cve_Municipio', columns = "Tipo de delito", values='delitos').reset_index()

df_homicies_15_19_wide['Cve_Municipio'] = df_homicies_15_19_wide['Cve_Municipio'].astype(str).str.zfill(5)

df_homicies_15_19_wide["Homicidio"].sum()

dict_variables = {"Cve_Municipio": "CVEGEO",
              "Extorsion":"EXT",
              "Homicidio":"HOMIC", 
              "Narcomenudeo":"DRUG_DEAL",
              "Robo": "THEFT",
              "Secuestro": "KIDNA"} 


df_homicies_15_19_wide = df_homicies_15_19_wide.rename(columns = )

dataframes_homic = [df_cvegeo,
                    df_homicides_15_w_var_f,
                    df_homicides_16_w_var_f,
                    df_homicides_17_w_var_f,
                    df_homicides_18_w_var_f,
                    df_homicides_19_w_var_f]

df_homics = reduce(lambda  left,right: pd.merge(left,right,on=['CVEGEO'],
                                            how = 'outer'), dataframes_homic).fillna(0)


df_homics["homic_15_19"] = df_homics[["homic_15", "homic_16", "homic_17", "homic_18", "homic_19"]].sum(axis=1)/5

df_homics["homic_19"].sum()

df_homics_15_19 = df_homics[["CVEGEO", "homic_15_19"]]

df_homics_15_19["homic_15_19"] = df_homics_15_19["homic_15_19"].round()

df_homics_15_19

df_homics_15_19.to_excel('data/df_homics_15_19.xlsx', index = False)
df_homics_15_19.to_parquet('data/df_homics_15_19.parquet')
