==> analysis/analysis_old/07.5_wrangle_cars.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_cars_2019 = pd.read_excel("data-raw/vmrc_anual_tr_cifra_2019.xlsx")
df_cars_2019


df_cars_2019['CVE_ENT'] = df_cars_2019['ID_ENTIDAD'].astype(str).str.zfill(2)

df_cars_2019['CVE_MUN'] = df_cars_2019['ID_MUNICIPIO'].astype(str).str.zfill(3)

df_cars_2019['CVEGEO'] = df_cars_2019['CVE_ENT'] + df_cars_2019['CVE_MUN']

df_cars_2019['vehicles'] = df_cars_2019['AUTO_OFICIAL'] + df_cars_2019['AUTO_PUBLICO'] + df_cars_2019['AUTO_PARTICULAR'] + df_cars_2019['CAM_PAS_OFICIAL'] + df_cars_2019['CAM_PAS_PUBLICO'] + df_cars_2019['CAM_PAS_PARTICULAR'] + df_cars_2019['CYC_CARGA_OFICIAL'] + df_cars_2019['CYC_CARGA_PUBLICO'] + df_cars_2019['CYC_CARGA_PARTICULAR'] + df_cars_2019['MOTO_OFICIAL'] + df_cars_2019['MOTO_DE_ALQUILER'] + df_cars_2019['MOTO_PARTICULAR'] 



df_cars_2019_f = df_cars_2019[["CVEGEO", "vehicles"]]

df_cars_2019_f

dict_SEDENA = {"MILITAR_MUERTO_NUM": "deaths_militar","MILITAR_HERIDO_NUM":"injur_militar",
                       "AGRESOR_MUERTO_NUM":"deaths_aggressor", "AGRESOR_HERIDO_NUM":"injur_aggressor",
               "AGRESOR_DETENIDO_NUM":"arrest_aggressor", "CLAVE_MUNICIPIO": "CVEGEO", "enfren": "confront"} 

df_SEDENA_08_19_w = df_SEDENA_08_19[["year", "CLAVE_MUNICIPIO", "MILITAR_HERIDO_NUM", "MILITAR_MUERTO_NUM", "AGRESOR_MUERTO_NUM", "AGRESOR_HERIDO_NUM", "AGRESOR_DETENIDO_NUM", "enfren"]].rename(columns = dict_SEDENA)

df_SEDENA_08_19_w

df_SEDENA_08_19_w_group = df_SEDENA_08_19_w.groupby(['year', 'CVEGEO'], as_index=0).sum()

df_SEDENA_08_19_w_group

df_SEDENA_group_15_19 = df_SEDENA_08_19_w_group[(df_SEDENA_08_19_w_group['year']>=2015)]

df_SEDENA_group_15_19

df_SEDENA_group_15_19["mortality_index"] = df_SEDENA_group_15_19["deaths_aggressor"]/df_SEDENA_group_15_19["injur_aggressor"]

df_SEDENA_group_15_19_w = df_SEDENA_group_15_19.groupby(['CVEGEO'], as_index=0).mean()

df_SEDENA_group_15_19_w_f = df_SEDENA_group_15_19_w[["CVEGEO", "confront"]]

df_SEDENA_group_15_19_w_f["confront"] = df_SEDENA_group_15_19_w["confront"].round()

df_SEDENA_group_15_19_w_f

df_SEDENA_group_15_19_w_f.to_excel('data/df_SEDENA_15_19.xlsx', index = False)
df_SEDENA_group_15_19_w_f.to_parquet('data/df_SEDENA_15_19.parquet')
