==> analysis/analysis_old/07.7_wrangle_access_highway.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

%cd "Documents/GitHub/pmx_gasoline_theft_prediction/"

df_highway = pd.read_csv("data-raw/Accesibilidad muy baja o baja a carretera pavimentada a nivel municipal, 2020.csv")
df_highway_w = df_highway.drop(['clave_entidad'], axis=1)



df_highway_w

dict_var_highway = {"clave_municipal": "CVEGEO",
              "poblacion_2020":"pob_2020",
                 "pob_acce_muy_baja_o_baja":"pob_highway",
                 "porc_pob_acce_muy_baja_o_baja":"porc_highway"}


df_highway_w = df_highway_w.rename(columns = dict_var_highway)

df_highway_w['pob_2020']= df_highway_w['pob_2020'].str.replace(',',"").astype(int)
df_highway_w['pob_highway']= df_highway_w['pob_highway'].str.replace(',',"").astype(int)
df_highway_w['CVEGEO'] = df_highway_w['CVEGEO'].astype(str).str.zfill(5)
df_highway_w

df_highway_w.to_excel('data/df_highway_access.xlsx', index = False)
df_highway_w.to_parquet('data/df_highway_access.parquet')
