==> analysis/analysis_old/09_Final_data_cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from functools import reduce

from google.colab import drive
drive.mount('/content/drive')

%cd 'Documents/GitHub/pmx_gasoline_theft_prediction/'

pwd

df_thefts_PMX_complete_ori = pd.read_excel("data/df_PMX_thefts_15_19_pp_final.xlsx", converters={'CVEGEO' : str})


df_thefts_PMX_complete_ori.columns

variables_to_maintain = ['CVEGEO','sum_intersects', 'Num_thefts', 
                         'COUNT','price_15_19', 'POB_TOT', 'confront', 
                         'IRS', 'homic_rate', 'pipeline', 'SBASC', 
                         'POB_6_16_NEB', 'POB_16_EBI', 'NO_SS', 
                         'OVPT', 'NO_SANIT', 'OVSAE', 'NO_DREN', 
                         'NO_ELECT', 'NO_LAVA', 'No_REFRI'
                         ]

df_thefts_PMX = df_thefts_PMX_complete_ori[variables_to_maintain].rename( columns = {'CVEGEO': 'Mun_code',
                                 'sum_intersects' : 'Pipe_intersec', 
                                 'Num_thefts' : 'Thefts_cases', 
                                 'COUNT' : 'Gas_stations',
                                 'price_15_19' : 'Gas_price', 
                                 'POB_TOT' : 'Population', 
                                 'confront' : 'Confront_events',
                                 'IRS' : 'IRS', 
                                 'homic_rate' : 'Homic_rate', 
                                 'pipeline' : 'Susceptible'})
                                 

df_thefts_PMX

df_thefts_PMX['Gas_stations'] = round(df_thefts_PMX['Gas_stations'] / df_thefts_PMX['Population'] * 100000, 2)
df_thefts_PMX['Gas_price'] = round(df_thefts_PMX['Gas_price'],2)
df_thefts_PMX['Homic_rate'] = round(df_thefts_PMX['Homic_rate'],2)
df_thefts_PMX['IRS'] = round(df_thefts_PMX['IRS'],4)
df_thefts_PMX['Thefts_cases'] = df_thefts_PMX['Thefts_cases'].fillna(0)
df_thefts_PMX_cleaned = df_thefts_PMX.drop(['Thefts_cases'], axis = 1, inplace = False)
df_thefts_PMX_cleaned

df_thefts_PMX_cleaned.head()
df_vehicles_final

df_vehicles_final = pd.read_excel('data/df_vehicles_final.xlsx')

df_thefts_PMX_final_2019_spat = pd.read_excel('data/df_thefts_PMX_final_2019.xlsx')


df_thefts_PMX_final_2019_spat_f = df_thefts_PMX_final_2019_spat[["Mun_code", "porc_highway", "CARR-KM", "LINCARR", "RED-KM", "LINRED", "DENSCARR", "DENSRED"]]

df_thefts_PMX_final_2019_spat_f['Mun_code'] = df_thefts_PMX_final_2019_spat_f['Mun_code'].astype(str).str.zfill(5)


df_spat_final = df_thefts_PMX_final_2019_spat_f
df_vehicles_final['Mun_code'] = df_vehicles_final['CVEGEO'].astype(str).str.zfill(5)


df_spat_final

dataframes = [df_thefts_PMX_cleaned, df_spat_final, df_vehicles_final]


df_thefts_PMX_cleaned_spat = reduce(lambda  left, right: pd.merge(left,right, on=['Mun_code'],
                                            how = 'left'), dataframes)#.fillna(0)

df_thefts_PMX_cleaned_spat = df_thefts_PMX_cleaned_spat.fillna(0)

df_thefts_PMX_cleaned_spat["vehicles_rate_15_19"] = round(df_thefts_PMX_cleaned_spat['vehicles_15_19'] / df_thefts_PMX_cleaned_spat['Population'] * 100000, 2)

df_thefts_PMX_cleaned_spat

df_thefts_PMX_cleaned.to_excel('data/df_thefts_PMX_final_v2.xlsx', index = False)

df_thefts_PMX_cleaned

df_thefts_PMX_cleaned_spat.to_excel('data/df_thefts_PMX_cleaned_spat.xlsx', index = False)
